Move aperture shapes to separate modules
